Made aws-chalice tag builtin to config
Removed the global utils function to inject the aws-chalice tag
into the user configured tags and made it part of the config object
as users cannot override this anyways.
